requirements.txt
deepgram-sdk
moviepy==2.0.0.dev2
imageio==2.25.1
pysrt==1.1.2
flask
openai
pyaudio
opencv-python
Flask-WTF
WTForms
Werkzeug
